+++ js/util.js
@@ -1,0 +1,39 @@
+// Фунция для поиска случайного числа в заданном интервале включительно. Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random#%D0%BF%D0%BE%D0%BB%D1%83%D1%87%D0%B5%D0%BD%D0%B8%D0%B5_%D1%81%D0%BB%D1%83%D1%87%D0%B0%D0%B9%D0%BD%D0%BE%D0%B3%D0%BE_%D1%87%D0%B8%D1%81%D0%BB%D0%B0_%D0%B2_%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%BD%D0%BE%D0%BC_%D0%B8%D0%BD%D1%82%D0%B5%D1%80%D0%B2%D0%B0%D0%BB%D0%B5
+const getRandomNumber = (min, max) => {
+  if (min > max) {
+    return 0;
+  }
+  let result = Math.random() * (max - min + 1) + min;
+  return result > max ? max : result;
+}
+
+// Функция, возвращающая случайное целое число из переданного диапазона включительно. Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random#%D0%BF%D0%BE%D0%BB%D1%83%D1%87%D0%B5%D0%BD%D0%B8%D0%B5_%D1%81%D0%BB%D1%83%D1%87%D0%B0%D0%B9%D0%BD%D0%BE%D0%B3%D0%BE_%D1%86%D0%B5%D0%BB%D0%BE%D0%B3%D0%BE_%D1%87%D0%B8%D1%81%D0%BB%D0%B0_%D0%B2_%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%BD%D0%BE%D0%BC_%D0%B8%D0%BD%D1%82%D0%B5%D1%80%D0%B2%D0%B0%D0%BB%D0%B5_%D0%B2%D0%BA%D0%BB%D1%8E%D1%87%D0%B8%D1%82%D0%B5%D0%BB%D1%8C%D0%BD%D0%BE
+const getRandomInteger = (min, max) => {
+  return Math.floor(getRandomNumber(min, max));
+}
+
+// Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно. Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Number/toFixed
+const getCoordinates = (min, max, digits) => {
+  if (min > max) {
+    return '0';
+  }
+  return getRandomNumber(min, max).toFixed(digits);
+}
+
+// Получаем случайный элемент массива:
+const getRandomArrayElement = (elements) => {
+  return elements[getRandomInteger(0, elements.length - 1)];
+};
+
+// Получаем уникальный массив из count элементов:
+const getUniqueArray = (allArr, count) => {
+  const newArr = [];
+  for (let i = 1; i <= count; i++) {
+    newArr.push(getRandomArrayElement(allArr));
+  }
+  const uniqueSet = new Set(newArr);
+  return Array.from(uniqueSet);
+}
+
+
+export {getRandomInteger, getCoordinates, getRandomArrayElement, getUniqueArray};
